Change schema and add hash password on db

+++ server/database/schema.ts
@@ -1,7 +1,7 @@
-import { integer, pgTable, varchar, boolean } from "drizzle-orm/pg-core";
+import { smallserial, pgTable, varchar, boolean } from "drizzle-orm/pg-core";
 
 export const user = pgTable("user", {
-  id: integer("id").primaryKey(),
+  id: smallserial("id").primaryKey(),
   username: varchar("username", { length: 255 }).notNull(),
   password: varchar("password", { length: 255 }).notNull(),
   isAdmin: boolean("is_admin").notNull().default(true),
